+++ src/failure.rs
@@ -11,4 +11,5 @@ use {
 pub struct Failure {
     pub error_code: i32,
     pub output: CommandOutput,
+    pub suggest_backtrace: bool,
 }